requirements.txt
openai>=1.0.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
pypdf>=3.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
